refactor: Consolidate imports from schema files for consistency and clarity
