+++ components/block-messages.tsx
@@ -1,6 +1,6 @@
 import { PreviewMessage } from './message'
 import { useScrollToBottom } from './use-scroll-to-bottom'
-import { Vote } from '@/lib/db/schema/vote'
+import { Vote } from '@/lib/db/schema'
 import { ChatRequestOptions, CreateMessage, Message } from 'ai'
 import { memo } from 'react'
 import equal from 'fast-deep-equal'
