+++ components/chat.tsx
@@ -6,7 +6,7 @@ import { useEffect, useState } from 'react'
 import useSWR, { useSWRConfig } from 'swr'
 
 import { ChatHeader } from '@/components/chat-header'
-import type { Vote } from '@/lib/db/schema/vote'
+import type { Vote } from '@/lib/db/schema'
 import { fetcher } from '@/lib/utils'
 
 import { Block, type UIBlock } from './block'
