+++ components/code-editor.tsx
@@ -6,7 +6,7 @@ import { python } from '@codemirror/lang-python'
 import { oneDark } from '@codemirror/theme-one-dark'
 import { basicSetup } from 'codemirror'
 import React, { memo, useEffect, useRef } from 'react'
-import { Suggestion } from '@/lib/db/schema/suggestion'
+import { Suggestion } from '@/lib/db/schema'
 
 type EditorProps = {
   content: string
