+++ components/data-stream-handler.tsx
@@ -3,7 +3,7 @@
 import { useChat } from 'ai/react'
 import { useEffect, useRef } from 'react'
 import { BlockKind } from './block'
-import { Suggestion } from '@/lib/db/schema/suggestion'
+import { Suggestion } from '@/lib/db/schema'
 import { initialBlockData, useBlock } from '@/hooks/use-block'
 import { useUserMessageId } from '@/hooks/use-user-message-id'
 import { cx } from 'class-variance-authority'
