+++ components/document-preview.tsx
@@ -11,7 +11,7 @@ import {
 import { UIBlock } from './block'
 import { FileIcon, FullscreenIcon, LoaderIcon } from './icons'
 import { cn, fetcher } from '@/lib/utils'
-import { Document } from '@/lib/db/schema/document'
+import { Document } from '@/lib/db/schema'
 import { InlineDocumentSkeleton } from './document-skeleton'
 import useSWR from 'swr'
 import { Editor } from './editor'
