+++ components/message-actions.tsx
@@ -3,8 +3,7 @@ import { toast } from 'sonner'
 import { useSWRConfig } from 'swr'
 import { useCopyToClipboard } from 'usehooks-ts'
 
-import type { Vote } from '@/lib/db/schema/vote'
-import { getMessageIdFromAnnotations } from '@/lib/utils'
+import type { Vote } from '@/lib/db/schema'
 
 import { CopyIcon, ThumbDownIcon, ThumbUpIcon } from './icons'
 import { Button } from './ui/button'
