+++ components/message.tsx
@@ -5,7 +5,7 @@ import cx from 'classnames'
 import { AnimatePresence, motion } from 'framer-motion'
 import { memo, useMemo, useState } from 'react'
 
-import type { Vote } from '@/lib/db/schema/vote'
+import type { Vote } from '@/lib/db/schema'
 
 import { DocumentToolCall, DocumentToolResult } from './document'
 import { PencilEditIcon, SparklesIcon } from './icons'
