+++ components/messages.tsx
@@ -3,7 +3,7 @@ import { PreviewMessage, ThinkingMessage } from './message'
 import { useScrollToBottom } from './use-scroll-to-bottom'
 import { Overview } from './overview'
 import { memo } from 'react'
-import { Vote } from '@/lib/db/schema/vote'
+import { Vote } from '@/lib/db/schema'
 import equal from 'fast-deep-equal'
 
 interface MessagesProps {
