+++ components/version-footer.tsx
@@ -6,7 +6,7 @@ import { useState } from 'react'
 import { useSWRConfig } from 'swr'
 import { useWindowSize } from 'usehooks-ts'
 
-import type { Document } from '@/lib/db/schema/document'
+import type { Document } from '@/lib/db/schema'
 import { getDocumentTimestampByIndex } from '@/lib/utils'
 
 import type { UIBlock } from './block'
