--- lib/db/schema/document.ts
@@ -1,33 +1,0 @@
-import type { InferSelectModel } from 'drizzle-orm'
-import {
-  pgTable,
-  varchar,
-  timestamp,
-  uuid,
-  text,
-  primaryKey,
-} from 'drizzle-orm/pg-core'
-import { user } from './user'
-
-export const document = pgTable(
-  'Document',
-  {
-    id: uuid('id').notNull().defaultRandom(),
-    createdAt: timestamp('createdAt').notNull(),
-    title: text('title').notNull(),
-    content: text('content'),
-    kind: varchar('text', { enum: ['text', 'code'] })
-      .notNull()
-      .default('text'),
-    userId: uuid('userId')
-      .notNull()
-      .references(() => user.id),
-  },
-  (table) => {
-    return {
-      pk: primaryKey({ columns: [table.id, table.createdAt] }),
-    }
-  }
-)
-
-export type Document = InferSelectModel<typeof document>
