+++ lib/editor/suggestions.tsx
@@ -9,7 +9,7 @@ import { createRoot } from 'react-dom/client'
 
 import { Suggestion as PreviewSuggestion } from '@/components/suggestion'
 
-import type { Suggestion } from '@/lib/db/schema/suggestion'
+import type { Suggestion } from '@/lib/db/schema'
 import { BlockKind } from '@/components/block'
 
 export interface UISuggestion extends Suggestion {
